+++ git_who/__init__.py
@@ -1,3 +1,3 @@
 """git-who: Find out who really knows your code."""
 
-__version__ = "0.6.0"
+__version__ = "0.7.0"